dsads

+++ src/components/List/index.jsx
@@ -1,12 +1,12 @@
-import React from 'react';
-import classNames from 'classnames';
-import axios from 'axios';
+import React from "react";
+import classNames from "classnames";
+import axios from "axios";
 
-import removeSvg from '../../assets/img/remove.svg';
+import removeSvg from "../../assets/img/remove.svg";
 
-import Badge from '../Badge';
+import Badge from "../Badge";
 
-import './List.scss';
+import "./List.scss";
 
 const List = ({
   items,
@@ -14,13 +14,15 @@ const List = ({
   onClick,
   onRemove,
   onClickItem,
-  activeItem
+  activeItem,
 }) => {
-  const removeList = item => {
-    if (window.confirm('Вы действительно хотите удалить список?')) {
-      axios.delete('http://localhost:3001/lists/' + item.id).then(() => {
-        onRemove(item.id);
-      });
+  const removeList = (item) => {
+    if (window.confirm("Вы действительно хотите удалить список?")) {
+      axios
+        .delete("https://narrow-gamy-chef.glitch.me/lists/" + item.id)
+        .then(() => {
+          onRemove(item.id);
+        });
     }
   };
 
@@ -32,7 +34,7 @@ const List = ({
           className={classNames(item.className, {
             active: item.active
               ? item.active
-              : activeItem && activeItem.id === item.id
+              : activeItem && activeItem.id === item.id,
           })}
           onClick={onClickItem ? () => onClickItem(item) : null}
         >
